% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{addFormulas}
\alias{addFormulas}
\title{Adding up two formulas, taking into account possible differing
 elements}
\usage{
addFormulas(formula1, formula2)
}
\arguments{
\item{formula1}{named numeric vector, example c(O = 2, C = 1)}

\item{formula2}{named numeric vector, example c(H = 2, S = 1)}
}
\value{
a named numeric vector (formula)
}
\description{
Adding up two formulas, taking into account possible differing
 elements
}
\examples{
addFormulas(waterFormula(), protonFormula())
addFormulas(waterFormula(), c(C=1, O=2))
}
