% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{addListFormulas}
\alias{addListFormulas}
\title{Add up a list of formulas}
\usage{
addListFormulas(formulas)
}
\arguments{
\item{formulas}{list of formulas}
}
\value{
a named numeric vector (formula)
}
\description{
Take a list of formulas and adds them all up
}
\examples{
addListFormulas(list(c(H = 2, O = 1),
                     c(H = 1),
                     c(H = 2, O = 1),
                     c(S = 1, O = 2)))
}
