% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{aminoAcidModifications}
\alias{aminoAcidModifications}
\title{Returns a pre-defined object which contains info on some common
 amino acid modifications}
\usage{
aminoAcidModifications()
}
\value{
An object of class modifications containing info on amino acid
 modifications
}
\description{
Returns a pre-defined object which contains info on some common
 amino acid modifications
}
\note{
the resulting modification table cannot be used immediately: there is
 two times a fixed modification for Cysteine amino acids. Remove one of them
 to prevent errors when using peptide calculations
}
\examples{
print(aminoAcidModifications)

}
