% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{calculate.Measured.mz}
\alias{calculate.Measured.mz}
\title{calculate.Measured.mz}
\usage{
calculate.Measured.mz(referenceMz, ppm)
}
\arguments{
\item{referenceMz}{reference mass or m/z in Da. Usually a theoretical mass
calculated from a formula}

\item{ppm}{deviation (in ppm) from the reference mass}
}
\value{
a numeric vector
}
\description{
calculates the mass (Da) when it deviates a certain ppm (part
 per million) from the reference mass (Da).
}
\examples{
calculate.Measured.mz(465.6025, 5)
calculate.Measured.mz(465.6025, 0)
calculate.Measured.mz(465.6025, -5)
}
