% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{calculate.ppm}
\alias{calculate.ppm}
\title{calculate.ppm}
\usage{
calculate.ppm(referenceMz, measuredMz)
}
\arguments{
\item{referenceMz}{reference mass or m/z in Da. Usually a theoretical mass
calculated from a formula}

\item{measuredMz}{measured mass or mz in Da.}
}
\value{
a numeric vector
}
\description{
calculates the mass deviation in ppm (parts per million) between
 the measured mass or m/z and a reference mass or m/z (both in Da).
}
\examples{
calculate.ppm(465.6025, 465.6028)
calculate.ppm(465.6025, 465.6025)
calculate.ppm(465.6025, 465.7025)
}
