% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{calculate.Reference.mz}
\alias{calculate.Reference.mz}
\title{calculate.Reference.mz}
\usage{
calculate.Reference.mz(measuredMz, ppm)
}
\arguments{
\item{measuredMz}{measured mass or m/z in Da.}

\item{ppm}{deviation (in ppm) that measuredMz differs from the reference mass
to be calculated}
}
\value{
a numeric vector
}
\description{
calculate the reference mass (Da) from a maesured mass (Da)
 which deviates a certain amount (ppm). Somewhat odd calculation, that is
 usually not needed in the lab, but is sometimes needed in theoretical
 situations or in programming.
}
\examples{
calculate.ppm(465.6025, 465.6028)
calculate.Measured.mz(465.6025, 0.644)
calculate.Reference.mz(465.6028, 0.644)
}
