% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{digest}
\alias{digest}
\title{Digests a sequence and returns}
\usage{
digest(sequence, enzyme = "trypsin", missed = 0)
}
\arguments{
\item{sequence}{character vector representing the amino acid sequence to be
digested. Note: the letters in sequence will be changed to upper case.}

\item{enzyme}{character string specifying the enzyme to be used for the
digestion. Default is 'trypsin'. Other options are 'trypsin.strict', 'pepsin',
'chymotrypsin' and 'chymotrypsin.strict'}

\item{missed}{integer vector: the maximum number of allowed missed cleavages}
}
\value{
data.frame with the columns 'peptide', 'start', 'stop' and 'mc'
 (missed cleavages)
}
\description{
Digests a sequence and returns
}
\note{
This function is an modified version of the Digest function found in
 the package 'OrgMassSpecR'
}
