% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{electronFormula}
\alias{electronFormula}
\title{generates a pre-defined formula for electron}
\usage{
electronFormula()
}
\value{
a named numeric vector (formula)
}
\description{
generates a pre-defined formula for electron
}
\note{
this is used for calculations
}
\examples{
print(electronFormula())
}
