% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{elements}
\alias{elements}
\title{R6 Class representing a set of elements}
\description{
Every element inside the object has a name, letter and a mass. The first 2
 can be any length of string, mass should be a numeric
}
\details{
Note: this class is meant to be used for elements.

Warning: all elements inside this object should be unique (names & shorts,
 not mass)
}
\examples{
randomElements <- elements$new(shorts = c("X1","X2","X3"),
                               names  = c("Secret Element 1",
                                          "Secret Element 2",
                                          "Secret Element 1"),
                               mass   = c(301, 312, 323))
                                                   
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{number}}{retrieve the number of elements present in the object, read only}

\item{\code{names}}{to access the names of the elements in the object}

\item{\code{shorts}}{to access the shorts of the elements in the object}

\item{\code{mass}}{to access the masses of the elements in the object}

\item{\code{table}}{retrieves all info on the elements in data.frame format, read only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elements-new}{\code{elements$new()}}
\item \href{#method-elements-addElement}{\code{elements$addElement()}}
\item \href{#method-elements-print}{\code{elements$print()}}
\item \href{#method-elements-getMass}{\code{elements$getMass()}}
\item \href{#method-elements-clone}{\code{elements$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elements-new"></a>}}
\if{latex}{\out{\hypertarget{method-elements-new}{}}}
\subsection{Method \code{new()}}{
creates a new elements object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elements$new(shorts, names, mass)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shorts}}{character vector specifying the short names for the elements, eg Hg for Mercury}

\item{\code{names}}{character vector specifying the names of the elements}

\item{\code{mass}}{numeric vector specifying the masses of the elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'elements' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elements-addElement"></a>}}
\if{latex}{\out{\hypertarget{method-elements-addElement}{}}}
\subsection{Method \code{addElement()}}{
adds one or more elements to the object. Elements to be added must have unique
 names and shorts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elements$addElement(shorts, names, mass)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shorts}}{character vector specifying the short names for the elements, eg Hg for Mercury}

\item{\code{names}}{character vector specifying the names of the elements}

\item{\code{mass}}{numeric vector specifying the masses of the elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elements-print"></a>}}
\if{latex}{\out{\hypertarget{method-elements-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints a table of the chemicals with columns letter, name & short
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elements$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elements-getMass"></a>}}
\if{latex}{\out{\hypertarget{method-elements-getMass}{}}}
\subsection{Method \code{getMass()}}{
Retrieves the mass of one of the elements in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elements$getMass(which1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which1}}{specifies which element should be retrieved. Which number (row number in
the chemicals table), name or short as a character vector. The way this is set up,
it doesn't matter whether capital or non-capital letters are used, since all are
converted to upper case before comparing with what's in the elements table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elements-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elements-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elements$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
