% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{elementsAverage}
\alias{elementsAverage}
\title{generates a pre-defined object which contains info on elements,
 mass values are average masses (weighted mean mass of elements based on
 their natural occurence)}
\usage{
elementsAverage()
}
\value{
an elements object
}
\description{
generates a pre-defined object which contains info on elements,
 mass values are average masses (weighted mean mass of elements based on
 their natural occurence)
}
\note{
electron is not really meant to be used in formulas, but is needed to
 calculate mass & m/z of ions
}
\examples{
print(elementsAverage())
}
