% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{elementsInFormula}
\alias{elementsInFormula}
\title{elementsInFormula}
\usage{
elementsInFormula(formula, removeZero = FALSE)
}
\arguments{
\item{formula}{named numeric vector, example c(O = 2, C = 1)}

\item{removeZero}{logical flag on how to deal with elements which are zero}
}
\value{
character vector
}
\description{
retrieves the names of the elements in a formula
}
\examples{
glucose = c(C=6, H=12, O=6, S=0)
elementsInFormula(glucose)
elementsInFormula(glucose, removeZero = TRUE)
}
