% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{elementsInFormulas}
\alias{elementsInFormulas}
\title{Combine elements present in 2 separate formulas}
\usage{
elementsInFormulas(formula1, formula2, decrease = FALSE)
}
\arguments{
\item{formula1}{named numeric vector, example c(O = 2, C = 1)}

\item{formula2}{named numeric vector, example c(H = 2, S = 1)}

\item{decrease}{logical flag on how to sort, default = FALSE: increasing}
}
\value{
a character vector
}
\description{
especially important when formula1 contains elements that
 formula2 does not contain and vice versa, Note: also sorts the result
}
\examples{
elementsInFormulas(c(O = 2, C = 1),
                   c(H = 2, S = 1))
}
