% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{elementsMonoisotopic}
\alias{elementsMonoisotopic}
\title{generates a pre-defined R6 elements object which contains info on
 elements, mass values are mono isotopic}
\usage{
elementsMonoisotopic()
}
\value{
an elements object
}
\description{
generates a pre-defined R6 elements object which contains info on
 elements, mass values are mono isotopic
}
\note{
this object is (by default) used in all chemical calculations in this
 package

electron is not really meant to be used in formulas, but is needed to
 calculate mass & m/z of ions
}
\examples{
print(elementsMonoisotopic())
}
