% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{emptyFormula}
\alias{emptyFormula}
\title{generates an empty pre-defined formula}
\usage{
emptyFormula()
}
\value{
a named numeric vector (formula)
}
\description{
generates an empty pre-defined formula
}
\note{
this can be used for calculations and setup of unknowns
}
\examples{
print(emptyFormula())
}
