% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{formulaToMass}
\alias{formulaToMass}
\title{calculates the neutral mono-isotopic mass of a formula}
\usage{
formulaToMass(
  formula = NULL,
  removeNA = FALSE,
  elementsInfo = elementsMonoisotopic()
)
}
\arguments{
\item{formula}{named numeric vector, example c(O = 2, C = 1)}

\item{removeNA}{logical vector: what to do if any of the elements is NA. If
TRUE, then remove before calculation, if FALSE, then do not remove}

\item{elementsInfo}{elements masses to be used, needs to be of class
elements, default is elementsMonoisotopic()}
}
\value{
numeric vector
}
\description{
calculates the neutral mono-isotopic mass of a formula
}
\examples{
formulaToMass(c(H=2, O=1))
formulaToMass(c(H=2, O=1), elementsInfo = elementsAverage())
}
