% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{\%f-\%}
\alias{\%f-\%}
\title{custom operator for subtracting formulas from one another, to make
 calculating with formulas a little more clear}
\usage{
formula1 \%f-\% formula2
}
\arguments{
\item{formula1}{named numeric vector, example c(O = 2, C = 1); formula to be
subtracted from}

\item{formula2}{named numeric vector, example c(H = 2, S = 1); formula to
subtract}
}
\description{
custom operator for subtracting formulas from one another, to make
 calculating with formulas a little more clear
}
\examples{
c(H = 2, O = 1) \%f-\% c(H = 1)
c(H = 2, O = 1) \%f-\% c(S = 1, O = 2)

}
