% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{\%f+\%}
\alias{\%f+\%}
\title{custom operator for adding up formulas, to make calculating with
 formulas a little more clear}
\usage{
formula1 \%f+\% formula2
}
\arguments{
\item{formula1}{named numeric vector, example c(O = 2, C = 1)}

\item{formula2}{named numeric vector, example c(H = 2, S = 1)}
}
\description{
custom operator for adding up formulas, to make calculating with
 formulas a little more clear
}
\examples{
waterFormula() \%f+\% protonFormula()
waterFormula() \%f+\% c(C=1, O = 2)
c(H = 2, O = 1) \%f+\% c(S = 1, O = 2)

}
