% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{massToMz}
\alias{massToMz}
\title{Calculates the m/z value of a charged/adducted ion}
\usage{
massToMz(
  mass,
  adducts = 0,
  adductFormula = electronFormula(),
  adductCharge = -1,
  elementsInfo = elementsMonoisotopic()
)
}
\arguments{
\item{mass}{numeric vector, (neutral) mass of the molecule}

\item{adducts}{numeric vector, number of adducts 'attached to' or 'removed
from' the (originally neutral) molecule}

\item{adductFormula}{formula (named numeric vector) of the adduct}

\item{adductCharge}{numeric vector indicating the actual charge per adduct}

\item{elementsInfo}{elements masses to be used, needs to be of class
elements, default is elementsMonoisotopic()}
}
\value{
numeric vector
}
\description{
Calculates the m/z value of a charged/adducted ion
}
\examples{
# amino acid residue lysine + water
lysineMass <- formulaToMass(aminoAcidResidues()$getFormula("K") \%f+\% waterFormula())
lysineMass
# M+H+ : adducts =  1, adductFormula = protonFormula(), adductCharge = 1
# singly charged/protonated ion (ESI)
massToMz(lysineMass,
         adducts = 1,
         adductFormula = protonFormula(),
         adductCharge = 1)
# Doubly charged/protonated ion (ESI)
massToMz(lysineMass,
         adducts = 2,
         adductFormula = protonFormula(),
         adductCharge = 1)
# M-H- : single, negatively charged
massToMz(lysineMass,
         adducts = -1,
         adductFormula = protonFormula(),
         adductCharge = 1)
# M+ : singly positively charged (molecular) ion (EI)
massToMz(lysineMass,
         adducts = -1,
         adductFormula = electronFormula(),
         adductCharge = 1)
# M- : singly negatively charged (molecular) ion (EI)
massToMz(lysineMass,
         adducts = 1,
         adductFormula = electronFormula(),
         adductCharge = 1)
}
