% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{modifications}
\alias{modifications}
\title{R6 Class representing a set of modifications for the aminoacids in peptides}
\description{
Every modification inside the object has a name, position, fixed (flag),
 gain (formula), loss (formula) and category.

'name' is a character vector.

 Position is a character vector specifying the amino acids which always have
  the modification (fixed = TRUE) or can have the modification (fixed =
  FALSE). More than one amino acid can be specified, eg NQ (for  Asparagine &
  glutamine). Please note that currently the package does NOT support
  'exotic' amino acids (eg Selenocysteine) or 'combination' letters, such as
  'J' (Leucine or Isoleucine). For the C- and N-terminus, use 'C_Term' or
  'N_Term' for position.

 Gain and loss specify what is lost and/or gained when a amino acid is
  modified. For example Carbamidomethylation of Cysteine has both a loss
  formula c(H=1) and a gain formula c(C=2, H=4, N=1, O=1); obviously this
  could also be defined as: loss formula = emptyFormula(), gain formula =
  c(C=2, H=3, N=1, O=1).

  For the category field (character vector) there is no real 'rule' on how to
  classify modifications. I usually stick to the categorisation of Mascot or
  Sequest.
}
\note{
the logical vector fixed is very important. If TRUE, then a
 modification is considered to be always present, if FALSE then its presence
 is optional.
}
\examples{
aaModifications <- modifications$new()
aaModifications$addTable(
  tibble::tibble(
    name = c("Carbamidomethyl (C)",
             "Carboxymethyl (C)",
             "Oxidation (M)"),
    position = c("C","C","M"),
    fixed = c(TRUE,TRUE,FALSE),
    gain = list(c(C = 2, H = 4, N = 1, O = 1),
                c(C = 2, H = 3, N = 0, O = 2),
                c(C = 0, H = 0, N = 0, O = 1)),
    loss = list(c(protonFormula()),
                c(protonFormula()),
                c(emptyFormula())),
    category = c("Cys-state","Cys-state","Preparation Artefact")
  )
)
aaModifications
aaModifications$add(name = "Deamidation",
                    position = "NQ",
                    fixed = FALSE,
                    gain = c(O = 1, H = 1),
                    loss = c(N = 1, H =2),
                    category = "Preparation Artefact")
aaModifications
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{number}}{retrieve the number of modifications present in the object, read only}

\item{\code{fixed}}{retrieve a table of the fixed modifications in the modification table, read only}

\item{\code{variable}}{retrieve a table of the variable modifications in the modification table, read only}

\item{\code{table}}{to access the table of modifications directly}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-modifications-new}{\code{modifications$new()}}
\item \href{#method-modifications-print}{\code{modifications$print()}}
\item \href{#method-modifications-add}{\code{modifications$add()}}
\item \href{#method-modifications-addTable}{\code{modifications$addTable()}}
\item \href{#method-modifications-clone}{\code{modifications$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modifications-new"></a>}}
\if{latex}{\out{\hypertarget{method-modifications-new}{}}}
\subsection{Method \code{new()}}{
Create a new modifications object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{modifications$new(data = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{default = NA. If not NA, then should be a tibble with 6 columns: name, position,
fixed, gain, loss and category. This is checked, but the contents of each column are
not checked.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modifications-print"></a>}}
\if{latex}{\out{\hypertarget{method-modifications-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints a table of the modifications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{modifications$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modifications-add"></a>}}
\if{latex}{\out{\hypertarget{method-modifications-add}{}}}
\subsection{Method \code{add()}}{
Adds a single modification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{modifications$add(name, position, fixed, gain, loss, category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character vector}

\item{\code{position}}{character vector, should be a valid amino acid residue}

\item{\code{fixed}}{logical vector, specifies whether the modification is fixed (TRUE) or
dynamic (FALSE)}

\item{\code{gain}}{named numeric vector (formula) that specifies what (atoms) are gained
when a modification is applied to an amino acid}

\item{\code{loss}}{named numeric vector (formula) that specifies what (atoms) are lost
when a modification is applied to an amino acid}

\item{\code{category}}{character vector. Not rigidly defined: for user to be able to
select/filter etc which type of modifications to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modifications-addTable"></a>}}
\if{latex}{\out{\hypertarget{method-modifications-addTable}{}}}
\subsection{Method \code{addTable()}}{
Add (a set of) modifications via a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{modifications$addTable(data = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{default = NA. If not NA, then should be a tibble with 6 columns: name, position,
fixed, gain, loss and category. This is checked, but the contents of each column are
not checked.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modifications-clone"></a>}}
\if{latex}{\out{\hypertarget{method-modifications-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{modifications$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
