% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{mzToMass}
\alias{mzToMass}
\title{calculates the mass of the molecule in an ion}
\usage{
mzToMass(
  mz,
  adducts = 0,
  adductFormula = electronFormula(),
  adductCharge = -1,
  elementsInfo = elementsMonoisotopic()
)
}
\arguments{
\item{mz}{mass to charge ratio of the ion}

\item{adducts}{numeric vector, number of adducts 'attached to' or 'removed
from' the (originally neutral) molecule}

\item{adductFormula}{formula (named numeric vector) of the adduct}

\item{adductCharge}{numeric vector indicating the actual charge per adduct}

\item{elementsInfo}{elements masses to be used, needs to be of class
elements, default is elementsMonoisotopic()}
}
\value{
numeric vector
}
\description{
essentially the reverse of the \code{\link{massToMz}}
}
\examples{
massToMz(mass = 174.1117, adductFormula = c(e=1), adducts = 2, adductCharge = -1) |>
 mzToMass(adductFormula = c(e=1), adducts = 2, adductCharge = -1)
massToMz(mass = 174.1117, adductFormula = c(H=1), adducts = 1, adductCharge = 1) |>
 mzToMass(adductFormula = c(H=1), adducts = 1, adductCharge = 1)
}
