% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{peptideCount}
\alias{peptideCount}
\title{counts the occurence of a amino acid (sequence) in another amino acid
 sequence}
\usage{
peptideCount(
  thePeptide = NA,
  searchPeptide = NA,
  doNotSplice = TRUE,
  upper = TRUE
)
}
\arguments{
\item{thePeptide}{character vector, the peptide to be searched}

\item{searchPeptide}{character vector, the amino acid sequence to search for}

\item{doNotSplice}{if FALSE the all characters in the searchPeptide are
searched individually.If TRUE then the searchPeptide is searched as a whole.
Default = TRUE}

\item{upper}{convert both thePeptide & searchPeptides to uppercase before
searching}
}
\value{
numeric vector
}
\description{
counts the occurence of a amino acid (sequence) in another amino acid
 sequence
}
\examples{
peptideCount("SAMPLER", "P")
peptideCount("SAMPLER", "PLER", doNotSplice = TRUE)
peptideCount("SAMPLER", "PLER", doNotSplice = FALSE)
}
