% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{peptideFormula}
\alias{peptideFormula}
\title{peptideFormula}
\usage{
peptideFormula(peptide, aminoAcids = aminoAcidResidues())
}
\arguments{
\item{peptide}{character vector specifying the sequence of amino acids in a
peptide}

\item{aminoAcids}{R6 object of type 'chemicals' with the amino acid info,
default = aminoAcidResidues()}
}
\value{
a numeric vector
}
\description{
gives formula of a peptide string
}
\note{
does not check for non-amino acid letters, modifications cannot be
 specified
}
\examples{
peptideFormula("SAMPLER")
}
