% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{peptideFragments}
\alias{peptideFragments}
\title{Generates a pre-defined (incomplete) table of names of fragments for
 the ions resulting when fragmenting a peptide in MS}
\usage{
peptideFragments()
}
\value{
a data.frame of two columns: 'name' and 'series name' of fragments
}
\description{
Generates a pre-defined (incomplete) table of names of fragments for
 the ions resulting when fragmenting a peptide in MS
}
\note{
will possibly be removed in the future
}
