% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{peptideMassToMz}
\alias{peptideMassToMz}
\title{peptideMassToMz}
\usage{
peptideMassToMz(
  peptide,
  charge = 1,
  aminoAcids = aminoAcidResidues(),
  elementsInfo = elementsMonoisotopic()
)
}
\arguments{
\item{peptide}{character vector specifying the sequence of amino acids in a
peptide}

\item{charge}{numeric vector specifying the charge of the peptide ion}

\item{aminoAcids}{R6 object of type 'chemicals' with the amino acid info,
default = aminoAcidResidues()}

\item{elementsInfo}{R6 object of type 'elements' with the elements masses
info, default = elementsMonoisotopic()}
}
\value{
a numeric vector
}
\description{
gives ion m/z of a peptide string c(C=x,H=y,etc etc)
}
\note{
does not check for non-amino acid letters, modifications cannot be
 specified
}
\examples{
peptideMassToMz("SAMPLER")
peptideMassToMz("SAMPLER", charge = 2)
peptideMassToMz("SAMPLER", elementsInfo = elementsAverage())
}
