% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{peptideMzH}
\alias{peptideMzH}
\title{peptideMzH}
\usage{
peptideMzH(
  peptide,
  charge = 1,
  aminoAcids = aminoAcidResidues(),
  elementsInfo = elementsMonoisotopic()
)
}
\arguments{
\item{peptide}{character vector specifying the sequence of amino acids in a
peptide}

\item{charge}{numeric vector specifying the charge of the peptide ion}

\item{aminoAcids}{R6 object of type 'chemicals' with the amino acid info,
default = aminoAcidResidues()}

\item{elementsInfo}{R6 object of type 'elements' with the elements masses
info, default = elementsMonoisotopic()}
}
\value{
a numeric vector
}
\description{
gives ion m/z of the protonated peptide
}
\note{
does not check for non-amino acid letters, modifications cannot be
 specified
}
\examples{
peptideMzH("SAMPLER")
peptideMzH("SAMPLER", charge = 2)
peptideMzH("SAMPLER", elementsInfo = elementsAverage())
}
