% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\name{peptideToFormula}
\alias{peptideToFormula}
\title{peptideToFormula}
\usage{
peptideToFormula(peptide, aminoAcids = aminoAcidResidues())
}
\arguments{
\item{peptide}{character vector specifying the sequence of amino acids in a
peptide}

\item{aminoAcids}{R6 object of type 'chemicals' with the amino acid info,
default = aminoAcidResidues()}
}
\value{
a numeric vector
}
\description{
gives formula of a peptide string c(C=x,H=y,etc etc)
}
\note{
does not check for non-amino acid letters, modifications cannot be
 specified
}
\examples{
peptideToFormula("SAMPLER")
}
