% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{protonFormula}
\alias{protonFormula}
\title{generates a pre-defined formula for proton}
\usage{
protonFormula()
}
\value{
a named numeric vector (formula)
}
\description{
generates a pre-defined formula for proton
}
\examples{
print(protonFormula())
}
