% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{rcdkFormula}
\alias{rcdkFormula}
\title{translates an cdkFormula object to a 'regular' formula format}
\usage{
rcdkFormula(cdkformula)
}
\arguments{
\item{cdkformula}{an object of type rcdkFormula}
}
\value{
formula of format c(H=2, O=1)
}
\description{
translates an cdkFormula object to a 'regular' formula format
}
\note{
This function does not deal with the charge state which is possibly
 defined in the rcdkFormula object
}
\examples{
glucose <- rcdk::get.formula("C6H12O6")
rcdkFormula(glucose)
glucoseAdductIon <- rcdk::get.formula("C6H12O6Na1", charge = 1)
glucoseAdductIon
# to get to the same m/z value
glucoseAdductIon |> massSpectrometryR::rcdkFormula() |> formulaToMass() |> massToMz(adducts = -1)
}
