% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{removeZeros}
\alias{removeZeros}
\title{removeZeros}
\usage{
removeZeros(formula)
}
\arguments{
\item{formula}{named numeric vector, example c(O = 2, C = 1)}
}
\value{
named numeric vector (formula)
}
\description{
removes elements that have number zero
}
\examples{
glucose = c(O=6, H=12, C=6)
glucose \%f+\% emptyFormula()
removeZeros(glucose \%f+\% emptyFormula())
}
