% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{sortFormula}
\alias{sortFormula}
\title{sortFormula}
\usage{
sortFormula(formula, decrease = FALSE)
}
\arguments{
\item{formula}{named numeric vector, example c(O = 2, C = 1)}

\item{decrease}{logical flag on how to sort, default = FALSE: increasing}
}
\value{
named numeric vector (formula)
}
\description{
sorts the elements of a formula in alphabetical order
 (increasing/decreasing)
}
\examples{
glucose = c(O=6, H=12, C=6)
glucose
sortFormula(glucose)
}
