% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{stringFormula}
\alias{stringFormula}
\title{Translates a character vector formula, eg 'C6H12O6' to a regular
 formula c(C=6, H=12, O=6)}
\usage{
stringFormula(string)
}
\arguments{
\item{string}{character vector, format eg: 'C6H12O6'}
}
\value{
formula of format c(H=2, O=1)
}
\description{
Translates a character vector formula, eg 'C6H12O6' to a regular
 formula c(C=6, H=12, O=6)
}
\note{
it's imperative that every element has a number (count), otherwise this
 function is highly likely to malfunction and return NA
}
\examples{
stringFormula("H3O4P1")
stringFormula("C6H12O6")
}
