% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{subtractFormulas}
\alias{subtractFormulas}
\title{subtracting one formula from another, taking into account possible
 differing elements}
\usage{
subtractFormulas(formula1, formula2)
}
\arguments{
\item{formula1}{named numeric vector, example c(O = 2, C = 1); formula to be
subtracted from}

\item{formula2}{named numeric vector, example c(H = 2, S = 1); formula to
subtract}
}
\value{
a named numeric vector (formula)
}
\description{
subtracting one formula from another, taking into account possible
 differing elements
}
\note{
There are no checks for negative values!
}
\examples{
subtractFormulas(c(H = 2, O = 1), c(H = 1))
subtractFormulas(c(H = 2, O = 1), c(S = 1, O = 2))
}
