% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{validFormula}
\alias{validFormula}
\title{checks if formula is valid}
\usage{
validFormula(formula, string = FALSE)
}
\arguments{
\item{formula}{character vector or named numeric vector, representing the
formula to be checked}

\item{string}{logical vector specifying if the formula is a character vector
or not}
}
\value{
logical vector
}
\description{
checks if formula is valid
}
\note{
This is done via the check_chemform function from the package enviPat
}
\examples{
glucose <- c(C=6, H=12, O=6)
validFormula(glucose)
formulaString(glucose)
validFormula(formulaString(glucose), string = TRUE)
}
