% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{waterFormula}
\alias{waterFormula}
\title{generates a pre-defined formula for water}
\usage{
waterFormula()
}
\value{
a named numeric vector (formula)
}
\description{
generates a pre-defined formula for water
}
\examples{
print(waterFormula())
}
