% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{formulaString}
\alias{formulaString}
\title{Translates regular formula format into a character vector, eg
 C6H12O6}
\usage{
formulaString(formula, removeSingle = FALSE, useMarkdown = FALSE)
}
\arguments{
\item{formula}{named numeric vector, example c(O = 2, C = 1)}

\item{removeSingle}{if TRUE then for elements that are present in the
formula only a single time, the number (1) will not be included. Default is
FALSE. See also examples}

\item{useMarkdown}{default = FALSE. If TRUE, the it will use HTML/Markdown
codes <sub> in the formulas, which can be used with the library 'gt' to
generate 'proper' notation for chemical formulas (numbers in subscript)}
}
\value{
character vector
}
\description{
Translates regular formula format into a character vector, eg
 C6H12O6
}
\examples{
formulaString(c(C=6, H=12, O=6))
formulaString(c(H=3,O=4,P=1))
formulaString(c(H=3,O=4,P=1), removeSingle = TRUE)
formulaString(c(H=2, O=1))
formulaString(c(H=2, O=1), removeSingle = TRUE)
}
