% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{massToMzH}
\alias{massToMzH}
\title{Calculates the m/z value of a protonated ion (positive ESI)}
\usage{
massToMzH(mass, charge = 1, elementsInfo = elementsMonoisotopic())
}
\arguments{
\item{mass}{numeric vector, (neutral) mass of the molecule}

\item{charge}{charge state}

\item{elementsInfo}{elements masses to be used, needs to be of class
elements, default is elementsMonoisotopic()}
}
\value{
numeric vector
}
\description{
a wrapper around \code{\link{massToMz}} for positively charged,
 protonated ions in ESI
}
\examples{
# amino acid residue lysine + water
lysineMass <- formulaToMass(aminoAcidResidues()$getFormula("K") \%f+\% waterFormula())
lysineMass
# M+H+ : adducts =  1, adductFormula = protonFormula(), adductCharge = 1
# singly charged/protonated ion (ESI)
massToMz(lysineMass,
         adducts = 1,
         adductFormula = protonFormula(),
         adductCharge = 1)
massToMzH(lysineMass)
}
