% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{stringToFormula}
\alias{stringToFormula}
\title{Translates a character vector formula, eg 'C6H12O6' to a regular
 formula c(C=6, H=12, O=6)}
\usage{
stringToFormula(string)
}
\arguments{
\item{string}{character vector, format eg: 'C6H12O6'}
}
\value{
formula of format c(H=2, O=1)
}
\description{
Translates a character vector formula, eg 'C6H12O6' to a regular
 formula c(C=6, H=12, O=6)
}
\note{
this function is an improved version of stringFormula(). Not every elements
 with count 1 can have the number omitted. However, the function depends on 'correct'
 elements (first letter is uppercase, second letter is lowercase)
}
\examples{
stringToFormula("H3O4P1")
stringToFormula("C6H12O6")
stringToFormula("C6H5Br")
}
